Fix checkstyle issues, Improve README, Refactor storage.load() to improve abstraction and readability.
