+++ data/duke.txt
@@ -1,3 +1,4 @@
 T@1@watch anime@ @fun @
-D@0@finish episode @2021-05-21@ @
+D@1@finish episode @2021-05-21@ @
 T@0@watch lecture episode@ @ @
+T@0@visit Japan!!@ @YAY uwu??? @
