+++ src/main/java/duke/exceptions/DukeDataLoadException.java
@@ -1,7 +1,5 @@
 package duke.exceptions;
 
-import duke.Duke;
-
 public class DukeDataLoadException extends DukeExceptions {
 
     public DukeDataLoadException(String message) {
