+++ src/main/java/duke/ui/MainWindow.java
@@ -7,20 +7,15 @@
 
 package duke.ui;
 
+import java.util.Objects;
+
+import duke.Duke;
 import javafx.fxml.FXML;
-import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-
-import duke.DialogBox;
-import duke.Duke;
-import duke.exceptions.DukeExceptions;
-
-import java.io.IOException;
-import java.util.Objects;
 
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
@@ -34,8 +29,6 @@ public class MainWindow extends AnchorPane {
     private VBox dialogContainer;
     @FXML
     private TextField userInput;
-    @FXML
-    private Button sendButton;
 
     private final Image userImage = new Image(Objects.requireNonNull(
             this.getClass().getResourceAsStream("/images/DaUser.png")));
@@ -70,7 +63,6 @@ public class MainWindow extends AnchorPane {
     @FXML
     private void handleUserInput() {
         String input = userInput.getText();
-        System.out.println(input);
         assert input != null : "The user input cannot be null.";
         String response = duke.getResponse(input);
 
@@ -81,3 +73,4 @@ public class MainWindow extends AnchorPane {
         userInput.clear();
     }
 }
+
